+++ src/main/java/com/aqualen/vsu/trueSkill/FactorGraphs/FactorGraphLayer.java
@@ -1,7 +1,12 @@
 package com.aqualen.vsu.trueSkill.FactorGraphs;
 
 import com.aqualen.vsu.entity.User;
+import com.aqualen.vsu.trueSkill.FactorGraphs.schedule.Schedule;
+import com.aqualen.vsu.trueSkill.FactorGraphs.schedule.ScheduleSequence;
+import com.aqualen.vsu.trueSkill.FactorGraphs.variable.KeyedVariable;
+import com.aqualen.vsu.trueSkill.GameInfo;
 import com.aqualen.vsu.trueSkill.GaussianDistribution;
+import com.aqualen.vsu.trueSkill.Player;
 import lombok.Getter;
 
 import java.util.ArrayList;
@@ -30,6 +35,8 @@ public abstract class FactorGraphLayer {
         inputVariables = (List) value;
     }
 
+    public abstract void buildLayer(GameInfo gameInfo, List<Player> players);
+
     public Object getRawOutputVariablesGroups()
     {
         return outputVariables;
@@ -38,5 +45,13 @@ public abstract class FactorGraphLayer {
     protected void addLayerFactor(Factor factor)
     {
         localFactors.add(factor);
+    }
+
+    protected Schedule<GaussianDistribution> scheduleSequence(
+    List<Schedule<GaussianDistribution>> itemsToSequence,
+    String nameFormat,
+    Object[]... args){
+        String formattedName = String.format(nameFormat, args);
+        return new ScheduleSequence(formattedName, itemsToSequence);
     }
 }
