+++ src/main/java/com/aqualen/vsu/trueSkill/Factors/GaussianLikelihoodFactor.java
@@ -1,7 +1,7 @@
 package com.aqualen.vsu.trueSkill.Factors;
 
 import com.aqualen.vsu.trueSkill.FactorGraphs.Message;
-import com.aqualen.vsu.trueSkill.FactorGraphs.Variable;
+import com.aqualen.vsu.trueSkill.FactorGraphs.variable.Variable;
 import com.aqualen.vsu.trueSkill.GaussianDistribution;
 
 import java.util.List;
